--- mep3_simulation/webots_data/protos/arm.protos
@@ -1,102 +1,0 @@
-Solid {
-  translation 0 0 0.08
-  scale 0.001 0.001 0.001
-  children [
-    Shape {
-      appearance USE WALL_APPEARANCE
-      geometry Mesh {
-        url [
-          "/home/bici/foxy_ws/src/mep3/mep3_simulation/webots_data/protos/arm_base.obj"
-        ]
-      }
-    }
-    HingeJoint {
-      jointParameters HingeJointParameters {
-        axis 0 0 1
-      }
-      endPoint Solid {
-        translation 0 82 0
-        rotation 0 0 1 0
-        children [
-          DEF ARM_MID_SHAPE Shape {
-            appearance USE WALL_APPEARANCE
-            geometry Mesh {
-              url [
-                "/home/bici/foxy_ws/src/mep3/mep3_simulation/webots_data/protos/arm_mid1.obj"
-              ]
-            }
-          }
-          HingeJoint {
-            jointParameters HingeJointParameters {
-              axis 0 0 1
-            }
-            endPoint Solid {
-              translation 0 82 0
-              rotation 0 0 1 0
-              children [
-                USE ARM_MID_SHAPE
-                HingeJoint {
-                  jointParameters HingeJointParameters {
-                    axis 0 0 1
-                  }
-                  endPoint Solid {
-                    rotation 0 0 1 0
-                    children [
-                      Shape {
-                        appearance USE WALL_APPEARANCE
-                        geometry Mesh {
-                          url [
-                            "/home/bici/foxy_ws/src/mep3/mep3_simulation/webots_data/protos/arm_vaccuum.obj"
-                          ]
-                        }
-                      }
-                    ]
-                    boundingObject Transform {
-                      translation 0 28 0
-                      children [
-                        Box {
-                          size 28 80 48
-                        }
-                      ]
-                    }
-                    physics Physics {
-                    }
-                  }
-                }
-              ]
-              boundingObject DEF TEST Transform {
-                translation 0 -40 0
-                children [
-                  Box {
-                    size 34 114 50
-                  }
-                ]
-              }
-              physics Physics {
-              }
-            }
-          }
-        ]
-        boundingObject DEF TEST2 Transform {
-          translation 0 -40 0
-          children [
-            Box {
-              size 34 114 50
-            }
-          ]
-        }
-        physics Physics {
-        }
-      }
-    }
-  ]
-  name "arm"
-  boundingObject Transform {
-    translation 0 -10.5 0
-    children [
-      Box {
-        size 36 57 50
-      }
-    ]
-  }
-}
