+++ mep3_simulation/webots_data/worlds/assets/.gitignore
@@ -3,3 +3,5 @@
 !table.png
 !samples/
 !samples/*.png
+!robot_arm_mesh/
+!robot_arm_mesh/*.obj
